cephadm: move monitoring class to a new file
