+++ src/cephadm/cephadmlib/daemons/__init__.py
@@ -4,6 +4,7 @@ from .ingress import HAproxy, Keepalived
 from .nvmeof import CephNvmeof
 from .iscsi import CephIscsi
 from .nfs import NFSGanesha
+from .monitoring import Monitoring
 
 __all__ = [
     'CephIscsi',
@@ -11,6 +12,7 @@ __all__ = [
     'CustomContainer',
     'HAproxy',
     'Keepalived',
+    'Monitoring',
     'NFSGanesha',
     'Tracing',
 ]
